% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_cp.R
\name{plot_cp}
\alias{plot_cp}
\title{Plot mCPR Data and Estimates}
\usage{
plot_cp(dat, est, iso_code, CI = 95)
}
\arguments{
\item{dat}{A tibble containing observed mCPR values. Columns: iso, year, cp, division_numeric_code & start_date/end_date and is_in_union == Y}

\item{est}{A tibble containing mCPR estimates. Columns: "Country or area", iso, Year, Median, U95, L95, U80, L80.}

\item{iso_code}{Country ISO code for filtering data.}

\item{CI}{Confidence intervals to be plotted, options: 95 (default), 80, or NA (no CI plotted).}
}
\value{
A ggplot object displaying the mCPR data and estimates, with or without confidence intervals.
}
\description{
Function plots modern contraceptive prevalence rates (mCPR) among married women from observed data and estimates.
}
\examples{
plot_cp(dat, est, iso_code = 4)
plot_cp(dat, est, iso_code = 404, CI = 80)

}
